test/Data/HealthCardIDTest.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package Data;

import org.junit.After;
import org.junit.AfterClass;
import org.junit.Before;
import org.junit.BeforeClass;
import org.junit.Test;
import static org.junit.Assert.*;

/**
 *
 * @author mfoulouyvesmarcel
 */
public class HealthCardIDTest {
    
    public HealthCardIDTest() {
    }
    
    @BeforeClass
    public static void setUpClass() {
    }
    
    @AfterClass
    public static void tearDownClass() {
    }
    
    @Before
    public void setUp() {
    }
    
    @After
    public void tearDown() {
    }

    /**
     * Test of getPersonalID method, of class HealthCardID.
     */
    @Test
    public void testGetPersonalID() {
        System.out.println("getPersonalID");
        HealthCardID instance = null;
        String expResult = "";
        String result = instance.getPersonalID();
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }

    /**
     * Test of equals method, of class HealthCardID.
     */
    @Test
    public void testEquals() {
        System.out.println("equals");
        Object o = null;
        HealthCardID instance = null;
        boolean expResult = false;
        boolean result = instance.equals(o);
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }

    /**
     * Test of hashCode method, of class HealthCardID.
     */
    @Test
    public void testHashCode() {
        System.out.println("hashCode");
        HealthCardID instance = null;
        int expResult = 0;
        int result = instance.hashCode();
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }

    /**
     * Test of toString method, of class HealthCardID.
     */
    @Test
    public void testToString() {
        System.out.println("toString");
        HealthCardID instance = null;
        String expResult = "";
        String result = instance.toString();
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }
    
}
